Primeiro Ano/1st Semester/SD/SD-PROJETO/backend/requirements.txt
aioice==0.10.1
aiortc==1.14.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
av==16.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
google-crc32c==1.7.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
mnemonic==0.21
platformdirs==4.5.0
pycparser==2.23
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
python-engineio==4.12.3
python-socketio==5.14.3
requests==2.32.5
simple-websocket==1.1.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
